feat/MGS-5727/update-magento into 2.x

See merge request magesuite/daily-deal!97

+++ Plugin/UpdateOfferInCart.php
@@ -57,7 +57,7 @@ class UpdateOfferInCart
                 continue;
             }
 
-            $qty = isset($data[$itemId]['qty']) ? (double)$data[$itemId]['qty'] : false;
+            $qty = isset($data[$itemId]['qty']) ? (float)$data[$itemId]['qty'] : false;
             $oldQty = $item->getQty();
 
             if (!$qty || $qty <= $oldQty) {
